refactor(core): switch to couchdb with simple http post request
